--- src/Header/SearchContainer/NameSearch/NameSearch.js
@@ -1,9 +1,0 @@
-const NameSearch = () => {
-  return (
-    <form>
-      <input type='text'/>
-    </form>
-  )
-}
-
-export default NameSearch
