--- src/Header/SearchContainer/SearchContainer.js
@@ -1,13 +1,0 @@
-import NameSearch from './NameSearch/NameSearch'
-import DropdownFilter from './DropdownFilter/DropdownFilter'
-
-const SearchContainer = () => {
-  return (
-    <section>
-      <NameSearch />
-      <DropdownFilter />
-    </section>
-  )
-}
-
-export default SearchContainer
