+++ src/HomeView/Header/Header.js
@@ -1,13 +1,12 @@
-import SearchContainer from './SearchContainer/SearchContainer'
 import ToggleView from './ToggleView/ToggleView'
-import '../images/floral-background.jpg'
+import '../../images/floral-background.jpg'
 import './Header.scss'
 
 const Header = ({ toggleView }) => {
   return (
     <header className='header'>
       <h1 className='title'>floristic</h1>
-      <SearchContainer />
+      
       <ToggleView toggleView={ toggleView }/>
     </header>
   )
