+++ src/apiCalls.js
@@ -26,9 +26,9 @@ export const getPlantsByLocation = async (slug) => {
   return data
 }
 
-// export const searchPlants = async (query) => {
-//   let url = `http://localhost:8080/https://trefle.io/api/v1/plants/search?q=${query}token=w76udTztX_89MySv3fO4fG41HD2yq9xhIRETq1KCXCg`
-//   const response = await fetch(url)
-//   const data = await response.json()
-//   return data
-// }
+export const searchPlants = async (query) => {
+  let url = `http://localhost:8080/https://trefle.io/api/v1/plants/search?q=${query}token=w76udTztX_89MySv3fO4fG41HD2yq9xhIRETq1KCXCg`
+  const response = await fetch(url)
+  const data = await response.json()
+  return data
+}
